requirements.txt
absl-py==2.1.0
addict==2.4.0
albumentations==1.0.3
altgraph==0.17.4
colorama==0.4.6
coloredlogs==15.0.1
configparser==7.0.0
cycler==0.12.1
flatbuffers==23.5.26
future==0.18.3
grad-cam==1.5.4
grpcio==1.60.1
humanfriendly==10.0
imageio==2.34.0
imgviz==1.7.5
importlib-metadata==7.0.1
joblib==1.3.2
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.3.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
onnx==1.15.0
onnxruntime==1.17.0
opencv-contrib-python==4.8.1.78
opencv-python==4.5.2.54
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.3
pefile==2023.2.7
Pillow==9.5.0
protobuf==4.25.3
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.2
PyMySQL==1.1.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2024.2
PyWavelets==1.4.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
scikit-image==0.19.3
scikit-learn==1.4.1.post1
scipy==1.10.1
shutils==0.1.0
six==1.16.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2024.2.12
torch==1.9.0+cu111
torchaudio==0.9.0
torchsummaryX==1.3.0
torchvision==0.10.0+cu111
tqdm==4.61.2
ttach==0.0.3
typing_extensions==4.8.0
tzdata==2024.2
Werkzeug==3.0.1
zipp==3.17.0